moved more functions to utils attempt batting average
